[feedback] Address the Ravin's suggestions
